requirements.txt
eth-brownie==1.18.1
pynvim==0.4.3
pyright==1.1.239
black==22.3.0
